+++ frontend/package.json
@@ -25,6 +25,7 @@
     "react-dom": "^18.2.0",
     "react-feather": "^2.0.10",
     "react-oidc-context": "^2.2.1",
+    "react-router-dom": "^6.15.0",
     "react-table": "^7.8.0"
   },
   "devDependencies": {
