+++ frontend/src/App.tsx
@@ -1,44 +1,35 @@
-import axios from "axios";
-import { useAuth } from "react-oidc-context";
-import { DefaultHero } from "./views/hero";
-import { Layout } from "./views/layout";
-import { GroupsView } from "./views/groups";
+import React from "react";
+import ReactDOM from "react-dom/client";
+import { AuthProvider } from "react-oidc-context";
+import "./index.css";
+import { onSigninCallback } from "./helpers/auth";
+import { QueryClientProvider, QueryClient } from "@tanstack/react-query";
+import { RouterProvider } from "react-router-dom";
+import { router } from "./routes";
 
-const App = () => {
-  const auth = useAuth();
+const queryClient = new QueryClient();
 
-  switch (auth.activeNavigator) {
-    case "signinSilent":
-      return <div>Signing you in...</div>;
-    case "signoutRedirect":
-      return <div>Signing you out...</div>;
-  }
-
-  if (auth.isLoading) {
-    return <p>Loading...</p>;
-  }
-
-  if (auth.error) {
-    return <p>Oops... {auth.error.message}</p>;
-  }
-
-  if (auth.isAuthenticated) {
-    axios.defaults.headers.common[
-      "Authorization"
-    ] = `Bearer ${auth.user?.access_token}`;
-
-    return (
-      <Layout auth={auth}>
-        <GroupsView />
-      </Layout>
-    );
-  }
-
-  return (
-    <Layout auth={auth}>
-      <DefaultHero auth={auth} />
-    </Layout>
-  );
+const configuration = {
+  client_id: "219919",
+  redirect_uri: "http://localhost:3000",
+  scope: "openid profile onlineweb4",
+  authority: "https://old.online.ntnu.no/openid",
+  metadataUrl:
+    "https://old.online.ntnu.no/openid/.well-known/openid-configuration",
+  silent_redirect_uri: "http://localhost:3000",
+  filterProtocolClaims: true,
+  loadUserInfo: true,
+  revokeTokensOnSignout: true,
+  post_logout_redirect_uri: "http://localhost:3000",
+  onSigninCallback: onSigninCallback,
 };
 
-export default App;
+ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
+  <React.StrictMode>
+    <AuthProvider {...configuration}>
+      <QueryClientProvider client={queryClient}>
+        <RouterProvider router={router} />
+      </QueryClientProvider>
+    </AuthProvider>
+  </React.StrictMode>
+);
